Adds Gitter API authentication support
The Gitter API follows Oauth 2.0 for authenticating users. The flow is
based on browser redirects and code exchange as noted in Oauth 2.0. In
addition to this, OkHttp has been used as a networking library for its
ease of use and flexibility and Butterfnife for reducing code
complexity.

A WebView client has been implemented which supports javascript and
loads the authentication page of Gitter, where Gitter prompts the user
to authenticate his/her self. After successfull authentication, Gitter
redirects the WebView to a a webpage as inidcated by redirect uri
specified during creation of the app on Gitter with a code paramter
whose value is stored and is sent along with other parameters such as
client_id etc, as a post request using OkHttp to a token generation page
of Gitter which cross checks the parameters and returns an access code.
The access code is stored in shared preferences for future uses.

The current Chat Activity has been deleted and has been replaced by a
Dashboard Activity which is redirected to by another starting activity
known as Splash Activity which is used for redirects. When no internet
and access_code is available Splash Activity redirects to NoNetwork
Activity.

In addition to the above, the style colours of the app has also been
changed.