Adds room display support
Defined the UI of the app. It now has a bottom navigation bar for easy
navigation. Implements fragments with respect to it. The main fragments
which this commits adds are RoomsFragment, DashboardFragment,
SettingsFragment of which DashboardFragment and SettingsFragment are
both currently empty.

Realm database has been implemented for listing and adding rooms.

The RoomsFragment has a RecyclerView which lists out the rooms data from
the Realm db.

OkHTTP has been used to query the github api for getting out rooms.

Signed-off-by: Abishek V Ashok <abishekvashok@gmail.com>