Adds message display support (#9)
Defined the UI for the room, adds onClick listener for each room_item
inside the RoomsFragment which opens a newly created new RoomFragment
which displays the messages down in a recycler view with the help of
MessagesAdapter. Messages will be taken from the Gitter Api and be
updated with the newly created MessagesTable which is a RealmObject.

Soft updation support has been added for messages and rooms which means
that any update made in the message or rooms would instantly reflect to
the user when he becomes online with much less lines of code by making
sorting done with the help of message's id. Based on the concept that
newly added messages will have a greater value than the previous ones.

In addition to the above, network calls will be timed based on user's
network connectivity and the app will request the user to connect to
internet only when there are no messages or rooms and internet access.

OkHTTP has been used as usual to query the Gitter Api passing along the
authorisation code.

Signed-off-by: Abishek V Ashok <abishekvashok@gmail.com>