Adds support for sending messages
Adds message sending support along with drafts. Adds a EditText inside
the RoomFragment along with a send button for sending the messages.
Messages once typed in will be saved as draft for the room in a new
column in the RoomsTable. Message will be sent if internet is vaialable
or the next time it becomes available using the SyncAdapter.

Messages will be synced using the SyncAdapter to the Gitter database
using a simple logic which will enable the messages to be automatically
synced to the Gitter Database using it's API.

Since, we are not dealing with any types of accounts, stub content
providers and account managers have been setup.

Messsage will be sent using OkHTTP and the Gitter API.

Signed-off-by: Abishek V Ashok <abishekvashok@gmail.com>